requirements.txt
flask
gunicorn
requests
Pillow
openai